pyproject.toml
[tool.poetry]
name = "mcp-rag-toolkit"
version = "0.1.0"
description = "A local RAG system with SQL + FAISS + Hugging Face"
authors = ["Aadithya Vishnu Sajeev <axdithya@gmail.com>"]
packages = [{ include = "mcp_rag_toolkit" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
mcp = {extras = ["cli"], version = "^1.9.4", allow-prereleases = true}
sentence-transformers = "^4.1.0"
faiss-cpu = "^1.11.0"
psycopg2-binary = "^2.9.10"
pandas = "^2.3.0"
fastmcp = {version = "^2.8.1", allow-prereleases = true}
python-docx = "^1.2.0"
pymupdf = "^1.26.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"